+++ lib/setup.c
@@ -1966,7 +1966,7 @@ static int LUKS2_check_encryption_sector(struct crypt_device *cd, uint64_t devic
 		uint64_t data_offset_bytes, uint32_t sector_size, bool modify_sector_size,
 		bool verify_data_area_alignment, uint32_t *ret_sector_size)
 {
-	uint32_t dmc_flags;
+	uint64_t dmc_flags;
 
 	assert(ret_sector_size);
 
@@ -3691,7 +3691,7 @@ int crypt_resize(struct crypt_device *cd, const char *name, uint64_t new_size)
 	struct crypt_dm_active_device dmdq, dmd = {};
 	struct dm_target *tgt = &dmdq.segment;
 	struct crypt_params_integrity params = {};
-	uint32_t supported_flags = 0, dmflags = 0;
+	uint64_t supported_flags = 0, dmflags = 0;
 	uint64_t old_size;
 	int r;
 
@@ -7436,7 +7436,7 @@ int crypt_keyslot_add_by_keyslot_context(struct crypt_device *cd,
  */
 int crypt_use_keyring_for_vk(struct crypt_device *cd)
 {
-	uint32_t dmc_flags;
+	uint64_t dmc_flags;
 
 	/* dm backend must be initialized */
 	if (!cd)
